Simplify wxAuiTabCtrl construction
This class is only supposed to be created by the library itself, so we
can change its constructor signature.

This is a (somewhat) backwards-incompatible change, but this is actually
desirable as we don't want to allow any code outside of the library to
create this class and continuing to do it will stop working correctly
after the changes in the upcoming commits.