Rewrite event generation and propagation in wxAuiNotebook
Do not use events to call wxAuiNotebook functions from wxAuiTabCtrl, as
this interfered with handling the events in the application code in
highly non-obvious ways. Just call wxAuiNotebook functions directly from
wxAuiTabCtrl code and stop handling wxEVT_AUI_XXX in wxAuiNotebook
itself.

Add wxAuiTabEventSource to avoid making wxAuiTabCtrl itself friend of
wxAuiNotebook: this allows to define the subset of private functions of
the latter that the former can call.

Also fix a couple of places where wrong indices, corresponding to the
position of the tabs in their tab control, were used instead of the
correct logical page indices.

Closes #25544.