Avoid collapsing wxDVC node if already done by event handler
If an event handler for wxEVT_DATAVIEW_SELECTION_CHANGED called
Collapse() itself (which is weird but allowed), Collapse() still tried
to collapse it again, which could result in the count of items becoming
negative and all sorts of other problems.

Avoid them by not doing anything if the item has been collapsed after
the selection change.

Closes #25631.