Use wp_pointer_warp_v1 if available in wxWindow::WarpPointer()
Add code for retrieving Wayland globals to get the interface of the
protocol allowing to control mouse position under Wayland.

Because this protocol requires providing a valid last enter event
serial, also handle "enter" event on Wayland pointer interface to get
this serial, as it seems to be impossible to retrieve it from GTK, even
though it already has it internally.

This, in turn, requires handling Wayland seat addition and removal, but
this doesn't seem to have much overhead, so it's probably not a problem
to do it even for the systems that don't support the pointer warp
protocol (which is, at the time of this writing, almost all of them,
although this will hopefully change in the future).

fixup