Fix test for .symver asm directive support in configure
We need to use it with the correct, i.e. existing, symbol name, which
means the mangled name actually used by the linker and not just the name
of the function as it appears in the source.

Fix the test to actually detect lack of support for .symver when using
clang with -flto.

This commit is the same as f2b844307d (Fix test for .symver asm
directive support in configure, 2025-05-25) in 3.2 branch.

See #25438.