Fix MSVS debug visualizer of wxStrings
Fix the visualizer to display full Unicode string.

Previously, only the first character was shown in recent versions of
Visual Studio. This was due to the use of "sb" format specifier meaning
an ANSI string.

Use condition expressions to cover both wchar_t and UTF-8 builds.

Correct the specifier for UTF-8 case to be "s8b" so that it interprets
and displays UTF-8 correctly.