Remove notifier from the old model in generic wxDataViewCtrl
Failure to do so could result in a crash later if the old model was not
destroyed but continued to exist independently of the control.

This also makes the code more natural and removes the need for the
comment explaining why we reset the notifier if we don't have any model.

Closes #25686.