Avoid using the style attribute in wxSVGFileDC
When an online SVG is opened in a browser, the browser might not want to process
the style attributes because there is no valid Content Security Policy (CSP).

Use the attributes directly inside the svg elements.
This passes the W3C validator, except for the 'white-space' attribute. So keep this one inside a style attribute.