Fix showing extra controls in wxFileDialog under macOS
These controls were not shown due to a forgotten colon in the test for
setAccessoryViewDisclosed support added in e765756555 (OSX file dialog
extensions (#2592), 2021-11-26).

Fix this simply by adding the colon, which is part of the method name,
but also avoid calling this method for save file dialogs which not only
don't need it, but result in an immediate application exit if it is
called on them.

See #25717.

(combined cherry pick from 426f745c6b069e9cd71a47a2272f5d6e07a9c6df and
32ea27c34f4b88eb6d407c223d0b91c5c884554b)