Fix compilation with old Wayland client protocol headers
Only enable Wayland support if WL_MARSHAL_FLAG_DESTROY is defined, as it
is used in the pointer-warp protocol header and it can't be compiled
without it.

It would be better to require high enough version of wayland-client but
Wayland documentation doesn't state in which version this constant, and
wl_proxy_marshal_flags() function, was added.