Initialize wxDCImpl members in their declarations
Make it more obvious that all members are initialized correctly.

This requires not using bit fields for boolean members any longer, as
they can only be initialized in their declarations with C++20, but it
doesn't seem like we really need to use bit fields here anyhow, wxDC
objects are not optimized for space anyhow and the program is never
going to have that many of them.

Also inline ctor and dtor for a tiny performance gain.