Use number of displayed, and not document, lines in the map code
We really care about the number of physical, i.e. "display", lines
appearing in the window and not the number of logical, or "document",
lines.

They are the same as long as no lines are either wrapped or folded but
become different as soon as either of them is used.