Fix check for extended locale functions in CMake
HAVE_LOCALE_T is misnamed, as it must only be defined not just when
locale_t type is available but when the functions using it, such as
strxxx_l(), are available.

Due to the confusion about the nature of this symbol, CMake didn't check
for it correctly and only checked for its existence, but musl (used by
e.g. Alpine) does define the type without providing any of the functions
using it, which resulted in build errors later.

Fix this by checking for both the type and the functions in CMake too,
just as we do in configure.

See #25749, #25750.

(cherry picked from commit 170c38631458dcef648c31b9d98dcc20a8c57f9c)