Reduce wxPropertyGrid gutter width
Using the actual expander icon size instead of hard-coded ones (see 229a9631)
causes the gutter around the icons to be too big.

In general the icon seems wide enough to not need a gutter. Only on Windows
without using the system theme (i.e. when the +/- expander is used) the icon
becomes to small. Calculate a gutter size that will increase the icon width to
a standard 16px width.