Don't put wxStatusBar fields into corners on macOS
macOS 26 (and to lesser extent, macOS 11+) uses rounded corners for
windows and status bar fields were rendered into the corner.

Improve the appearance by offsetting the fields are appropriately to
look better. Also add a way to customize the offset from user code,
because the default could be unsuitable when e.g. the window uses larger
corner radius (possible on macOS 26), or when user code wants to do
smarter placement of the fields to mimic Finder (which centers the
statusbar, but not under the entire window, but under the main area sans
the sidebar).