Sync map and document properly when lines are wrapped
Don't assume that display and document lines are the same any longer.

This complicates things a lot, notably while dragging because we can't
simply find the correct first line for the editor any longer but need to
search for it iteratively because the size of the thumb now depends on
its position, as different regions of the document may wrap differently.

Also synchronize the map with the document when it is repainted and not
resized because line wrapping is done in the background after the resize
and we don't have the correct mapping between the display and document
lines yet in the wxEVT_SIZE handler -- but we eventually get it for the
last wxEVT_PAINT provoked by it.