CMake: Remove duplicate find_package
Packages of third-party dependencies were searched twice. Once when adding them via wx_add_thirdparty_library to determine if they are available or of the built-in library should be used. And then again when including the third-party cmake file. Move all the logic into the third-party library cmake file.

This also removes the need to move the WebP variables into the parent scope.