Silence warning about non-virtual dtor on VC++ <=2017
Explicitly defining a destructor in wxMSWOwnerDrawnButtonBase causes
precisely the kind of compiler warnings (C4265) that the definition was
meant to suppress on GCC.

C4265 is not enabled by default in VC++, but if you have it enabled,
you'll get plenty of pointless warnings through anything that #includes
wx/msw/ownerdrawnbutton.h. This is also the only spot in wx headers that
triggers C4265.

The issue appears to be 3.2-specific.