Rationalize the new added "first pass" functions
Rename them to have a name actually meaning something to people not
familiar with the details of the wxSizer layout algorithm who don't even
know that it makes 2 passes.

Also call the new wxWindow function GetMinSizeUsingLayoutDirection(),
without the "Effective" part, as GetEffectiveMinSize() is supposed to be
just a trivial wrapper around GetMinSize() and should have never been
virtual and overridable in the first place, even if it was unfortunately
made so back in a20a357d92 (Make GetEffectiveMinSize virtual as not
being able to override the method, which the sizer system depends on, is
inflexible, 2009-01-30).