Fix nested markup conversion to NSAttributedString
wxMarkupToAttrStringBase implementation of markup parsing failed to
correctly apply nested attributes: in e.g. <i><b>foo</b></i>, only the
outer attribute was in effect.

This was due to applying - and overwriting - attributes to a span in
OnAttrEnd(), which is called first for the inner span, and then for the
outer one, overwriting all inner changes.

Instead, apply the currently effective attributes to text in OnText(),
similarly to how wxMarkupParserRenderOutput does it. As a side effect,
this also eliminates the need to apply default font to the entire range
first.

Closes #25864.