Fix use of wxGC with a scrolled window in drawing sample
A clip region set using SetDeviceClippingRegion() on a DC should not scroll
when the canvas we are drawing on scrolls.  The drawing sample behaves like
this if the DC is a wxDC instance but not if it is a wxGC instance (because
it is not properly prepared when the canvas is scrolled, i.e.: GetDeviceOrigin()
always returns (0, 0) for a wxGC). So this bug should be fixed by preparing
the correct DC that we are actually using for drawing.

Closes #25877.