Fix scrolling the editor while dragging the thumb
Find the candidate middle of the thumb directly from the first visible
line in the editor, we don't need to compute the first visible map line
(especially as we did it wrongly) for this.

Get rid of m_dragLastY which is not needed to get a reasonably accurate
initial estimation.