Don't scroll wxStyledTextCtrl used by wxStyledTextCtrlMiniMap
Scrolling the control using wxWindow::ScrollWindow() corrupted the thumb
drawn on top of it, as parts of it were scrolled away. There doesn't
seem to be any way to prevent this from happening, so make it possible
to avoid using ScrollWindow() completely by adding new SetCustomDrawn()
function and using it in wxStyledTextCtrlMiniMap.

This fixes visual thumb corruption under wxMSW when scrolling the map by
small increments, e.g. using mouse wheel or the scrollbar arrows.