Don't synchronize map position on resize to avoid unwanted jumps
Map position may change when its size changes, but this shouldn't result
in a change in the editor position, as it did before this change because
calling SyncMapPosition() from wxEVT_SIZE handler resulted in a
subsequent call to SyncEditPosition() when we received a V_SCROLL update
from Scintilla.

Stop doing this from this handler and instead remember the last "set"
(actually, "known") map position to avoid reacting to the corresponding
V_SCROLL change.

Note that just setting m_lastSetMapFirst but continuing to call
SyncMapPosition() from here didn't quite work because Scintilla seems to
not be in a consistent state yet when this event handler is executed
when maximizing the window containing the map and the map position could
pass by some small intermediate value (2 seen in the sample) before
finally jumping to 0, which prevented us from filtering this update out.

Updating map only from wxEVT_PAINT handler doesn't seem to result in any
visual problems and avoids this problem.