Don't hide editor vertical scrollbar when using minimap
This looks nicer but loses potentially useful functionality of the
editor scrollbar which may be used to scroll the editor even if the text
is short enough to fit entirely in the map, and so when the map doesn't
have any scrollbar at all.

Document the possibility of hiding it scrollbar for people who still
want to do it.