Make configure check for dl_iterate_phdr() more precise
Check that this function is declared in link.h and not just that it
exists, as we also use HAVE_DL_ITERATE_PHDR to guard link.h inclusion.

Normally it shouldn't matter as the only systems on which this function
is present are those using glibc which declares it in this header, but
it's still tidier to actually check for what we use.

Closes #25921.