Fix using outdated unwrapped label in wxStaticText
Since the changes of 65e8f2b7ca (Merge branch 'stattext-wrap',
2025-10-25), see #25925 and #25753, a wrong label could be displayed
after a sequence of calls to Wrap(), SetLabel() and Wrap() with a
different value again.

This was due to reusing outdated value of m_unwrappedLabel, even after
the actual label has changed, so the solution is to invalidate
m_unwrappedLabel in SetLabel(). Unfortunately, this results in some
rather ugly code in Wrap(), which now has to restore m_unwrappedLabel
after using it, but there doesn't seem to be any other way to do it
currently.

The ideal solution would be to separate setting the label at wx level
and the actually displayed label in different functions, but this
requires updating all ports and so won't be done for now.

See #25795.