Get rid of DEFINE_STD_WXCOLOUR_CONSTRUCTORS() macro
This macro was always ugly and got even worse after the recent changes
adding many overloaded constructors to it, so refactor wxColour
declaration to avoid it: define port-specific wxColourImpl and inherit
the real wxColour from it, which allows us to define all these ctors in
a single place only.

There should be no changes for the existing code.