Change initial values of some wxAuiToolBarItem fields
Don't set m_dropDown and m_sticky to true initially, this didn't make
much sense because most items don't have drop downs and are not sticky.

Also change the default m_minSize value from (0,0) to (-1,-1) as the
code in wxAuiToolBar always effectively initialized it with it.

This is a backwards-incompatible change, formally, but as application
code shouldn't ever create wxAuiToolBarItem objects (because it can't do
anything useful with them) but only receive them from wxAuiToolBar, and
because this commit doesn't change the values of the fields of these
objects, it shouldn't have any consequences in practice.