Fix crash in wxAuiTabCtrl due to use of invalid m_hoverButton
This pointer could become invalid if the page containing it was
destroyed (typically, although not necessarily, as the result of
clicking on this button itself) and using it when handling the next
WXEVT_MOTION event resulted in a crash.

Avoid the problem by removing this pointer and replacing its use with
FindHoverButton() that finds the (only) button in hover state.

Also add ResetHoverButton() helper which uses FindHoverButton() and
resets the state of the button and updates the window if necessary.

Closes #25959.