Fix mouse handling while moving over buttons in wxAuiTabCtrl
There were several problems in wxAuiTabCtrl: a button could become
highlighted when we were passing over it with the mouse button pressed
outside of it (which is wrong, as releasing the mouse button doesn't
activate the button in this case) and a button could also remain
highlighted if we released the mouse button not over it, as we wrongly
switched to tab dragging mode, which shouldn't be done while a button is
pressed.

Fix both of them by ensuring we never have a pressed button and
m_isDragging == true at the same time.