Use WIDE variant of win32 API functions
wxWidgets uses the WIDE variant of the win32 API. But when the ANSI version of
Windows.h is included before wxWidgets headers and wxNO_WIN32_W is defined,
the ANSI variant of the functions will be used.

This can cause build errors in headers that are (indirectly) included by the user.
Because the class/string types do not match the function type.

So use the WIDE variant explicitly to prevent these errors.