Fix creating bitmap compatible with HDC in wxMSW
Strangely enough, the code for creating bitmap compatible with the given
HDC never actually did this, but always created bitmap compatible with
the screen.

Even though this was done like this ever since 2cf45d6944 (modifications
for raw bitmap support and for using DIBs, 2003-03-22), it still seems
wrong, so do use the HDC provided to this function if it is valid and
only fall back on the screen HDC otherwise.

This should hopefully result in no visible changes to behaviour, but
makes things more logical and a tad more efficient as we avoid the
unnecessary GetDC()/ReleaseDC() calls now.