Use 24bpp backing store bitmap in wxMSW to speed up drawing on it
Avoid the overhead of converting the backing store bitmap to DIB and
back when drawing bitmaps with alpha on it in AlphaBlt(), which may be
very significant, by using 24bpp bitmap: as it doesn't have alpha, it
doesn't need to be adjusted after drawing on it.

This is ugly and it would be better to have a better implementation of
AlphaBlt(), but it's not clear how to do it, and for now this fixes a
real user-visible problem.

See #14403, #23841.

(cherry picked from commit b732393986e5a4d73c79db5446f3d4b7b138d370)