Fix initial scale of wxGLCanvas under EGL/Wayland
The scale of the canvas was set up correctly only once we received
"size-allocate" signal from GTK, but this doesn't necessarily happen
when the window is first shown, e.g. the cube sample doesn't get this
signal, at least when using Sway.

Ensure that we still use the correct scale by catching "draw" signal as
well and updating the scale if it has changed.

See #23733, #25465.