Fix initial scale of wxGLCanvas under EGL/Wayland
The scale of the canvas was set up correctly only once we received
"size-allocate" signal from GTK, but this doesn't necessarily happen
when the window is first shown and after its scale factor is actually
known.

Ensure that we use the correct scale by catching the notification about
its change too. This makes buffer scale factor correct from the very
beginning, without having to wait for a resize.

See #23733, #25465.