Improve drawing wxListCtrl gird lines in dark mode
The grid lines are erased when custom drawing the sub-items, so the grid lines need to be redrawn after drawing the item.
OnPaint() is not called after every custom draw (e.g. just hovering over a row will cause a custom draw but no paint),
so it has to be drawn in the custom paint handler.

Add a common function for drawing grid lines and reuse it on OnPaint().