Don't forcible reset item min size to (1,1) in wxAUI
This had been already done once back in 5c62cb6c7d (wxAUI: maintain
minimum size of panes across dock/undock., 2006-07-05) before being
added back in dec588bd6e (reverted an earlier patch that was causing
problems, 2006-07-08) but it's not clear at all what kind of problems
not doing this was causing, while it's very clear that doing this
results in problems due to the window minimum size not being respected.

Try (un?)reverting this one more time: if we run into problems due to
this commit and have to undo it, hopefully these problems will at least
be documented in a comment and/or commit message this time.