Remove useless check from wxNotebook::GetTabRect() in wxMSW
There is no need to check that the notebook is not empty because this is
already ensured by the precondition check for the page index validity:
no pages can be valid in an empty notebook.

No real changes, just simplify the code added in 065ff2d2d8 (Add
wxNotebook::GetTabRect() to retrieve the tab position, 2022-11-03).

See #22941.

This commit is best viewed ignoring whitespace-only changes.