Ensure that closed panes are not shown in docking toolbars later
After the changes of the previous commit, closed panes were not shown in
the docking toolbars immediately after closing them, but could still be
shown in them if the toolbar was created later.

Fix this by remembering that the pane was closed in its state field and
skipping over closed panes when adding panes to the dock later.