Allow minimized pane icon to be different from the normal one
Add wxAuiPaneInfo::IconMin() which can be used to set the icon to shown
when the pane is minimized.

This is useful especially in the case when normal icon is not set at
all, as the minimized pane button really must have an icon.

Show it in the sample using some completely random items (which is not
ideal, but there are no good looking standard icons that would make
sense here).