Fix OpenGL ES creation in EGL-based wxGLCanvas implementation
The existing code tried to use EGL_RENDERABLE_TYPE as context attribute,
but it is not supported there.

Use eglBindAPI() instead, which also should be called even when not
using ES. To make this work, add a new EGL-specific member to
wxGLContextAttrs: this is a bit ugly, but not more so than the existing
GLX-specific members and is much simpler than any alternative.

See #26012.

Co-authored-by: Timothy Lee <timothy.ty.lee@gmail.com>