Increase click area for +/- buttons in wxGenericTreeCtrl
The native control reacts to clicks on the whole area around the
expander button, not just the chevron/triangle itself, so mimic this in
the generic control too.

Enable this behaviour for wxGTK, in addition to wxOSX (it should
probably be enabled unconditionally, but wasn't tested in other ports).

See #26029.