Make state of items in wxCheckListBox accessible
The standard Windows LISTBOX class used by wxCheckListBox internally has
no concept of checked/unchecked state of items and so the information
about the item state can't be returned by the built-in implementation of
IAccessible for this standard control.

Provide a custom implementation of IAccessible for wxCheckListBox which
extends the standard one to return the items state too.

Closes #25948.

Closes #26055.