Update generic wxTreeCtrl state images when DPI changes
Ensure that wxTreeCtrl using only state images, without the normal ones,
is still updated correctly when the DPI changes: this didn't happen
previously because OnImagesChanged() was only called for the control
with normal images.

As changing HasImages() to return true if there are only state images is
almost certainly a bad idea for compatibility reasons (existing code may
assume that it can use normal images if it returns true), add a new
HasAnyImages() function and override it in wxTreeCtrl to account for the
state images too.

Closes #26059.